refactor(core): remove unused calculateAllFileMetrics function and rename file
Remove calculateAllFileMetrics function that was only used in tests and rename calculateAllFileMetrics.ts to calculateSelectiveFileMetrics.ts to match the remaining function. This simplifies the metrics calculation API by keeping only the selective file metrics functionality that is actually used in production code.
