+++ tests/core/metrics/calculateMetrics.test.ts
@@ -2,11 +2,8 @@ import { type Mock, describe, expect, it, vi } from 'vitest';
 import type { ProcessedFile } from '../../../src/core/file/fileTypes.js';
 import type { GitDiffResult } from '../../../src/core/git/gitDiffHandle.js';
 import { TokenCounter } from '../../../src/core/metrics/TokenCounter.js';
-import {
-  calculateAllFileMetrics,
-  calculateSelectiveFileMetrics,
-} from '../../../src/core/metrics/calculateAllFileMetrics.js';
 import { calculateMetrics } from '../../../src/core/metrics/calculateMetrics.js';
+import { calculateSelectiveFileMetrics } from '../../../src/core/metrics/calculateSelectiveFileMetrics.js';
 import type { RepomixProgressCallback } from '../../../src/shared/types.js';
 import { createMockConfig } from '../../testing/testUtils.js';
 
@@ -19,8 +16,7 @@ vi.mock('../../../src/core/metrics/TokenCounter.js', () => {
   };
 });
 vi.mock('../../../src/core/metrics/aggregateMetrics.js');
-vi.mock('../../../src/core/metrics/calculateAllFileMetrics.js', () => ({
-  calculateAllFileMetrics: vi.fn(),
+vi.mock('../../../src/core/metrics/calculateSelectiveFileMetrics.js', () => ({
   calculateSelectiveFileMetrics: vi.fn(),
 }));
 
@@ -59,7 +55,6 @@ describe('calculateMetrics', () => {
     const gitDiffResult: GitDiffResult | undefined = undefined;
 
     const result = await calculateMetrics(processedFiles, output, progressCallback, config, gitDiffResult, {
-      calculateAllFileMetrics,
       calculateSelectiveFileMetrics,
       calculateOutputMetrics: () => Promise.resolve(30),
       calculateGitDiffMetrics: () => Promise.resolve(0),
