+++ tests/core/metrics/calculateSelectiveFileMetrics.test.ts
@@ -1,9 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
 import type { ProcessedFile } from '../../../src/core/file/fileTypes.js';
-import {
-  calculateAllFileMetrics,
-  calculateSelectiveFileMetrics,
-} from '../../../src/core/metrics/calculateAllFileMetrics.js';
+import { calculateSelectiveFileMetrics } from '../../../src/core/metrics/calculateSelectiveFileMetrics.js';
 import type { FileMetricsTask } from '../../../src/core/metrics/workers/fileMetricsWorker.js';
 import fileMetricsWorker from '../../../src/core/metrics/workers/fileMetricsWorker.js';
 import type { RepomixProgressCallback } from '../../../src/shared/types.js';
@@ -23,24 +20,7 @@ const mockInitTaskRunner = <T, R>(numOfTasks: number, workerPath: string) => {
   };
 };
 
-describe('calculateAllFileMetrics', () => {
-  it('should calculate metrics for all files', async () => {
-    const processedFiles: ProcessedFile[] = [
-      { path: 'file1.txt', content: 'a'.repeat(100) },
-      { path: 'file2.txt', content: 'b'.repeat(200) },
-    ];
-    const progressCallback: RepomixProgressCallback = vi.fn();
-
-    const result = await calculateAllFileMetrics(processedFiles, 'o200k_base', progressCallback, {
-      initTaskRunner: mockInitTaskRunner,
-    });
-
-    expect(result).toEqual([
-      { path: 'file1.txt', charCount: 100, tokenCount: 13 },
-      { path: 'file2.txt', charCount: 200, tokenCount: 50 },
-    ]);
-  });
-
+describe('calculateSelectiveFileMetrics', () => {
   it('should calculate metrics for selective files only', async () => {
     const processedFiles: ProcessedFile[] = [
       { path: 'file1.txt', content: 'a'.repeat(100) },
